+++ src/player.h
@@ -11,6 +11,7 @@ class Player
         {MINING, {1, 0}}
     };
     std::array<std::optional<Object>, INVENTORY_SIZE> inventory;
+    std::array<std::optional<Tool>, NUM_TOOLS> toolbelt;
     PlayerState player_state = NONE;
     int inventory_occupancy = 0;
 
@@ -34,13 +35,13 @@ class Player
             return true;
         }
 
-        size_t expProgress(Skills skill) noexcept
+        size_t expProgress(Skills skill) const
         {
             //ceil(a/b) = (a + b - 1)/b for unsigned integers
-            int player_exp = player_skills[skill].second;
-            if(player_skills[skill].first > 1)
-                player_exp -= level_exp_mapping(player_skills[skill].first);
-            int required_exp = level_exp_mapping(player_skills[skill].first + 1) - level_exp_mapping(player_skills[skill].first);
+            int player_exp = player_skills.at(skill).second;
+            if(player_skills.at(skill).first > 1)
+                player_exp -= level_exp_mapping(player_skills.at(skill).first);
+            int required_exp = level_exp_mapping(player_skills.at(skill).first + 1) - level_exp_mapping(player_skills.at(skill).first);
             if (player_exp == 0 || required_exp == 0)
                 return 0;
             return (player_exp * PROGRESSBAR_PARTITIONS + required_exp - 1)/required_exp;
@@ -85,7 +86,7 @@ class Player
             return inventory_occupancy == INVENTORY_SIZE;
         }
 
-        const std::array<std::optional<Object>, INVENTORY_SIZE>& getInventory() const
+        const std::array<std::optional<Object>, INVENTORY_SIZE>& getInventory() const noexcept
         {
             return inventory;
         }
@@ -100,6 +101,10 @@ class Player
 
             for (auto& slot : inventory)
                 slot.reset();
+
+            for (auto& tool : toolbelt)
+                tool.reset();
+
             inventory_occupancy = 0;
             player_state = NONE;
         }